refactor: check for entity presence inside service

+++ fintech/src/main/java/ru/kslacker/fintech/service/CityWeatherServiceImpl.java
@@ -10,6 +10,8 @@ import ru.kslacker.fintech.dataaccess.repositories.api.WeatherTypeInfoRepository
 import ru.kslacker.fintech.dto.CreateCityDto;
 import ru.kslacker.fintech.dto.CreateWeatherDto;
 import ru.kslacker.fintech.dto.WeatherDto;
+import ru.kslacker.fintech.exceptions.CityAlreadyExistsException;
+import ru.kslacker.fintech.exceptions.CityNotFoundException;
 import ru.kslacker.fintech.mapping.WeatherMapper;
 import ru.kslacker.fintech.dataaccess.entities.City;
 import ru.kslacker.fintech.dataaccess.entities.Weather;
@@ -31,6 +33,10 @@ public class CityWeatherServiceImpl implements CityWeatherService {
 
     @Override
     public List<WeatherDto> getWeather(UUID cityId, LocalDate date) {
+        if (!cityRepository.existsById(cityId)) {
+            throw new CityNotFoundException(cityId);
+        }
+
         return weatherRepository
                 .getByCityIdAndDateTimeBetween(cityId, date.atStartOfDay(), date.plusDays(1).atStartOfDay())
                 .stream()
@@ -42,6 +48,10 @@ public class CityWeatherServiceImpl implements CityWeatherService {
     @Transactional
     public void createCity(UUID cityId, CreateCityDto createCityDto) {
         validator.validate(createCityDto);
+        if (cityRepository.existsById(cityId)) {
+            throw new CityAlreadyExistsException(cityId);
+        }
+
         City city = cityRepository.save(new City(cityId, createCityDto.name()));
         List<Weather> weather = createWeatherList(city, createCityDto.weather());
         weatherRepository.createForCity(cityId, weather);
@@ -50,6 +60,9 @@ public class CityWeatherServiceImpl implements CityWeatherService {
     @Override
     @Transactional
     public void updateWeather(UUID cityId, CreateWeatherDto createWeatherDto) {
+        if (!cityRepository.existsById(cityId)) {
+            throw new CityNotFoundException(cityId);
+        }
         City city = cityRepository.getById(cityId);
         Weather weather = createWeather(city, createWeatherDto);
         weatherRepository.updateWeather(weather);
@@ -58,6 +71,9 @@ public class CityWeatherServiceImpl implements CityWeatherService {
     @Override
     @Transactional
     public void deleteCity(UUID cityId) {
+        if (!cityRepository.existsById(cityId)) {
+            throw new CityNotFoundException(cityId);
+        }
         cityRepository.deleteById(cityId);
         weatherRepository.deleteByCityId(cityId);
     }
